Moved updating a user in the database to new Model/User class. Updated calls in class to call new User static methods
