+++ app/src/main/java/com/CS2340/shelterapp/Model/User.java
@@ -1,0 +1,80 @@
+package com.CS2340.shelterapp.Model;
+
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.database.DataSnapshot;
+import com.google.firebase.database.DatabaseError;
+import com.google.firebase.database.DatabaseReference;
+import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
+
+/**
+ * User class to update user values in the database
+ * Created by chandler on 3/26/18.
+ */
+
+public class User {
+
+    /**
+     * Update a given user's disabled attribute in the database
+     * @param newDisabled string to be put into database
+     * @param user user to be updated with new value
+     */
+    public static void updateDBUserDisabled(String newDisabled, DatabaseReference user) {
+        if (newDisabled == null || user == null) {
+            return;
+        }
+        user.addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                user.child("Disabled").setValue(newDisabled);
+            }
+            @Override
+            public void onCancelled(DatabaseError databaseError) {
+                System.out.print(databaseError.getMessage());
+            }
+        });
+    }
+
+    /**
+     * Update a currentUser's checkedIn attribute in the database
+     * @param newCheckedIn number to be put into database
+     */
+    public static void updateDBUserCheckedIn(int newCheckedIn) {
+        DatabaseReference userDB = FirebaseDatabase.getInstance().getReference().child("Users");
+        FirebaseUser currentUser = FirebaseAuth.getInstance().getCurrentUser();
+        DatabaseReference userRef = userDB.child(currentUser.getUid());
+
+        userRef.addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                userRef.child("Checked In").setValue(newCheckedIn);
+            }
+            @Override
+            public void onCancelled(DatabaseError databaseError) {
+                System.out.print(databaseError.getMessage());
+            }
+        });
+    }
+
+    /**
+     * Update a currentUser's bed attribute in the database
+     * @param newBeds number to be put into database
+     */
+    public static void updateDBUserBeds(int newBeds) {
+        DatabaseReference userDB = FirebaseDatabase.getInstance().getReference().child("Users");
+        FirebaseUser currentUser = FirebaseAuth.getInstance().getCurrentUser();
+        DatabaseReference userRef = userDB.child(currentUser.getUid());
+
+        userRef.addListenerForSingleValueEvent(new ValueEventListener() {
+            @Override
+            public void onDataChange(DataSnapshot dataSnapshot) {
+                userRef.child("Beds").setValue(newBeds);
+            }
+            @Override
+            public void onCancelled(DatabaseError databaseError) {
+                System.out.print(databaseError.getMessage());
+            }
+        });
+    }
+}
